requirements.txt
numpy
pandas
opencv-python
tqdm
pillow
torch
torchvision